requirements.txt
orjson==3.10.7
PyYAML==6.0.2
requests==2.32.3
httpx==0.27.0
tenacity==8.5.0
pydantic==2.8.2
jsonschema==4.23.0
gspread==6.1.2
google-auth==2.33.0
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1

